requirements.txt
pandas
numpy
yfinance
scikit-learn
vectorbt
matplotlib
pyyaml
